+++ E.M.P/employManagementPlatform/src/main/java/ac/yuhan/domain/Pos.java
@@ -9,6 +9,6 @@ import lombok.Data;
 @Entity
 public class Pos {
 	@Id
-	private int p_num;
+	private Long p_num;
 	private String p_name;
 }
