+++ E.M.P/employManagementPlatform/src/main/java/ac/yuhan/persistence/DeptRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import ac.yuhan.domain.Dept;
 
-public interface DeptRepository extends JpaRepository<Dept, Integer> {
+public interface DeptRepository extends JpaRepository<Dept, Long> {
 
 }
