+++ E.M.P/employManagementPlatform/src/main/java/ac/yuhan/persistence/PosRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import ac.yuhan.domain.Pos;
 
-public interface PosRepository extends JpaRepository<Pos, Integer> {
+public interface PosRepository extends JpaRepository<Pos, Long> {
 
 }
